# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /manufacturing/orders/{id}/test-results:
    get:
      operationId: getOrderTestResults
      tags: [manufacturing]
      summary: Get all test results for an order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Test results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestResult"
  /manufacturing/orders/{id}/test-results/{testType}:
    put:
      operationId: upsertTestResult
      tags: [manufacturing]
      summary: Upsert a test result (capacity|charge_discharge|protection|internal_resistance)
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: testType
          in: path
          required: true
          schema:
            type: string
            enum: [capacity, charge_discharge, protection, internal_resistance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestResultInput"
      responses:
        "200":
          description: Upserted test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"

  # ── Manufacturing — QC Approval ──────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TestResult:
      type: object
      required: [id, productionOrderId, testType, operatorName, result, testData, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        productionOrderId: { type: string, format: uuid }
        testType: { type: string, enum: [capacity, charge_discharge, protection, internal_resistance] }
        testEquipmentId: { type: ["string", "null"], format: uuid }
        testEquipmentName: { type: ["string", "null"] }
        operatorName: { type: string }
        startedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        result: { type: string, enum: [pass, fail, warning] }
        testData: { type: object, additionalProperties: true }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TestResultInput:
      type: object
      required: [operatorName, result, testData]
      properties:
        testEquipmentId: { type: ["string", "null"], format: uuid }
        testEquipmentName: { type: ["string", "null"] }
        operatorName: { type: string }
        startedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        result: { type: string, enum: [pass, fail, warning] }
        testData: { type: object, additionalProperties: true }
        notes: { type: ["string", "null"] }
...